# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Sistema de Passagens",
...
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8044",
            "description": "Servidor API de Produção"
        }
    ],
    "paths": {
...
        "/api/baggages": {
            "get": {
                "tags": [
                    "Baggages"
                ],
                "summary": "Listar todas as bagagens",
                "operationId": "54fbdb9903cd8a7d7f7e4ff353fe4cde",
                "responses": {
                    "200": {
                        "description": "Lista de bagagens"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Baggages"
                ],
                "summary": "Criar uma nova bagagem",
                "operationId": "ec575514c7f8493d70d492ddf935e3e2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ticket_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bagagem criada com sucesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/baggages/{id}": {
            "get": {
                "tags": [
                    "Baggages"
                ],
                "summary": "Mostrar uma bagagem específica",
                "operationId": "924b662dc789127de92f5c8b387856c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da bagagem"
                    },
                    "404": {
                        "description": "Bagagem não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Baggages"
                ],
                "summary": "Atualizar uma bagagem",
                "operationId": "ad4e699dc512c4c75c824742a7401f60",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ticket_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bagagem atualizada com sucesso"
                    },
                    "404": {
                        "description": "Bagagem não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Baggages"
                ],
                "summary": "Deletar uma bagagem",
                "operationId": "5ce0b7f3268bddf047d29f7017295e05",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Bagagem deletada com sucesso"
                    },
                    "404": {
                        "description": "Bagagem não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/baggages/label": {
            "post": {
                "tags": [
                    "Baggages"
                ],
                "summary": "Emitir etiqueta da bagagem",
                "operationId": "97b93a776feca779ae780ca0bb681973",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ticket_number": {
                                        "type": "string",
                                        "example": "ticket-number-123"
                                    },
                                    "baggage_number": {
                                        "type": "string",
                                        "example": "baggage-number-456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Etiqueta da bagagem emitida com sucesso"
                    },
                    "404": {
                        "description": "Bagagem ou ticket não encontrado"
                    },
                    "400": {
                        "description": "Erro na emissão da etiqueta da bagagem"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a Bearer token para autenticação",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}